# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Battery Data Intelligence Engine",
    "description": "Intelligence API backing the admin dashboard. Protected routes verify Clerk session JWTs via JWKS; see app/api/deps.py.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/companies/{company_id}/facilities": {
      "get": {
        "tags": [
          "companies"
        ],
        "summary": "Get Company Facilities",
        "operationId": "get_company_facilities_api_v1_companies__company_id__facilities_get",
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FacilityRead"
                  },
                  "title": "Response Get Company Facilities Api V1 Companies  Company Id  Facilities Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/facilities": {
      "get": {
        "tags": [
          "facilities"
        ],
        "summary": "List Facilities",
        "operationId": "list_facilities_api_v1_facilities_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "company_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Company Id"
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          },
          {
            "name": "facility_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Facility Type"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_FacilityRead_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/facilities/{facility_id}/notes": {
      "get": {
        "tags": [
          "facilities"
        ],
        "summary": "List Facility Notes",
        "operationId": "list_facility_notes_api_v1_facilities__facility_id__notes_get",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Facility Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnalystNoteRead"
                  },
                  "title": "Response List Facility Notes Api V1 Facilities  Facility Id  Notes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "facilities"
        ],
        "summary": "Create Facility Note",
        "operationId": "create_facility_note_api_v1_facilities__facility_id__notes_post",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Facility Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalystNoteCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalystNoteRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/facilities/{facility_id}/verified": {
      "patch": {
        "tags": [
          "facilities"
        ],
        "summary": "Set Facility Verified",
        "operationId": "set_facility_verified_api_v1_facilities__facility_id__verified_patch",
        "parameters": [
          {
            "name": "facility_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Facility Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifiedUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifiedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalystNoteCreate": {
        "properties": {
          "note_type": {
            "type": "string",
            "enum": [
              "data_error",
              "missing_data",
              "outdated",
              "other"
            ],
            "title": "Note Type"
          },
          "note_text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 3,
            "title": "Note Text"
          }
        },
        "type": "object",
        "required": [
          "note_type",
          "note_text"
        ],
        "title": "AnalystNoteCreate",
        "description": "Request body for per-entity POST \u2026/notes endpoints.\n\nClients send only ``note_type`` and ``note_text``.  The route handler\ninjects ``entity_type`` and ``entity_id`` from the URL path before writing\nto ``analyst_notes``."
      },
      "AnalystNoteRead": {
        "properties": {
          "note_type": {
            "type": "string",
            "enum": [
              "data_error",
              "missing_data",
              "outdated",
              "other"
            ],
            "title": "Note Type"
          },
          "note_text": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 3,
            "title": "Note Text"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "entity_type": {
            "type": "string",
            "enum": [
              "company",
              "company_material_exposure",
              "company_supply_relationship",
              "company_regulation_exposure",
              "company_vehicle_model",
              "material",
              "hs_code_material_mappings",
              "regulation",
              "risk_event",
              "facility",
              "battery_chemistry"
            ],
            "title": "Entity Type"
          },
          "entity_id": {
            "type": "string",
            "title": "Entity Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "note_type",
          "note_text",
          "id",
          "entity_type",
          "entity_id",
          "created_at",
          "updated_at"
        ],
        "title": "AnalystNoteRead"
      },
...
      "FacilityRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "facility_type": {
            "type": "string",
            "title": "Facility Type"
          },
          "country": {
            "type": "string",
            "title": "Country"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "city": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "City"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "capacity_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Capacity Notes"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "verified": {
            "type": "boolean",
            "title": "Verified",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "facility_type",
          "country",
          "status"
        ],
        "title": "FacilityRead"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedResponse_FacilityRead_": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/FacilityRead"
            },
            "type": "array",
            "title": "Data"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          }
        },
        "type": "object",
        "required": [
          "data",
          "total",
          "page",
          "limit"
        ],
        "title": "PaginatedResponse[FacilityRead]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VerifiedResponse": {
        "properties": {
          "verified": {
            "type": "boolean",
            "title": "Verified"
          }
        },
        "type": "object",
        "required": [
          "verified"
        ],
        "title": "VerifiedResponse",
        "description": "Minimal response returned by all PATCH \u2026/verified endpoints."
      },
      "VerifiedUpdate": {
        "properties": {
          "verified": {
            "type": "boolean",
            "title": "Verified"
          }
        },
        "type": "object",
        "required": [
          "verified"
        ],
        "title": "VerifiedUpdate",
        "description": "Request body for PATCH \u2026/verified endpoints."
      }
    }
  }
}